Various CRUD improvements

+++ app/Http/Controllers/RoleController.php
@@ -8,6 +8,7 @@ use App\Models\RolePermission;
 use App\Models\User;
 use App\Services\SystemAuthorities;
 use Exception;
+use Faker\Provider\Uuid;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -20,28 +21,31 @@ class RoleController extends Controller
     public function getRoles()
     {
         if (!Gate::allows(SystemAuthorities::$authorities['view_role'])) {
-            return response()->json(['Message' => 'Not allowed to view roles: '], 500);
+            return response()->json(['message' => 'Not allowed to view roles: '], 500);
         }
-
-        $roles = Role::select(
-            "roles.name as role_name",
-            "roles.updated_at as updated_at",
-            "roles.meta as meta",
-            "roles.description as description",
-            "roles.uuid as uuid"
-        );
-
+        $roles = Role::where('deleted_at', null)->get();
+        if ($roles == null) {
+            return response()->json(['message' => 'Roles not found. '], 404);
+        }
+        foreach ($roles as $role) {
+            $role->permissions = $role->permissions()->get()->pluck('name','uuid');
+        }
         return  $roles;
     }
 
     public function getRole(Request $request)
     {
         if (!Gate::allows(SystemAuthorities::$authorities['view_role'])) {
-            return response()->json(['Message' => 'Not allowed to view role: '], 500);
+            return response()->json(['message' => 'Not allowed to view role: '], 500);
         }
-        $role = Role::find($request->id);
+        // $role = Role::find($request->id);
+        if ($request->uuid) {
+            $role = Role::where('uuid', $request->uuid)->first();
+        } else {
+            $role = Role::find($request->id);
+        }
         if($role == null){
-            return response()->json(['Message' => 'Role not found: '], 404);
+            return response()->json(['message' => 'Role not found. '], 404);
         }
         return  $role;
     }
@@ -49,80 +53,92 @@ class RoleController extends Controller
     public function createRole(Request $request)
     {
         if (!Gate::allows(SystemAuthorities::$authorities['add_role'])) {
-            return response()->json(['Message' => 'Not allowed to create roles: '], 500);
+            return response()->json(['message' => 'Not allowed to create roles: '], 500);
         }
         try {
-            $user = Auth::user();
+            //validate
+            $request->validate([
+                'name' => 'required',
+                'description' => 'required',
+            ]);
+            
             $role = new Role([
+                'uuid' => Uuid::uuid(),
                 'name' => $request->name,
                 'description' => $request->description,
-                'meta' => $request->meta,
+                'meta' => $request->meta ?? json_decode('{}'),
             ]);
             $role->save();
 
-            $permissions = $request->permission;
-            foreach ($permissions as $permission) {
-                $rolePermission = new RolePermission([
-                    'role' => $role->id,
-                    'permission' => $permission,
-                ]);
-                $rolePermission->save();
+            $permissions = $request->permissions;
+            if($permissions != null){
+                foreach ($permissions as $permission) {
+                    $rolePermission = new RolePermission([
+                        'uuid' => Uuid::uuid(),
+                        'role' => $role->uuid,
+                        'permission' => $permission,
+                    ]);
+                    $rolePermission->save();
+                }
             }
 
-            return response()->json(['Message' => 'Created successfully'], 200);
+            return response()->json(['message' => 'Created successfully'], 200);
         } catch (Exception $ex) {
 
-            return ['Error' => '500', 'Message' => 'Could not save role ' . $ex->getMessage()];
+            return ['Error' => '500', 'message' => 'Could not save role ' . $ex->getMessage()];
         }
     }
 
     public function deleteRole(Request $request)
     {
         if (!Gate::allows(SystemAuthorities::$authorities['delete_role'])) {
-            return response()->json(['Message' => 'Not allowed to delete roles: '], 500);
+            return response()->json(['message' => 'Not allowed to delete roles: '], 500);
         }
         try {
 
             $role = Role::find($request->id);
             $users =  User::where('role', '=', $request->id)->first();
             if (count($users) != 0) {
-                return response()->json(['Message' => 'Can\'t delete Role. Role is attached to users'], 500);
+                return response()->json(['message' => 'Can\'t delete Role. Role is attached to users'], 500);
             } else {
                 // DB::delete('delete from permissions where uuid in(select permission from role_permissions where role=?))', [$request->id]);
                 DB::delete('delete from role_permissions where role=?)', [$request->id]);
                 $role->delete();
             }
-            return response()->json(['Message' => 'Deleted successfully'], 200);
+            return response()->json(['message' => 'Deleted successfully'], 200);
         } catch (Exception $ex) {
-            return response()->json(['Message' => 'Delete failed.  Error code' . $ex->getMessage()], 500);
+            return response()->json(['message' => 'Delete failed.  Error code' . $ex->getMessage()], 500);
         }
     }
 
     public function updateRole(Request $request)
     {
         if (!Gate::allows(SystemAuthorities::$authorities['edit_role'])) {
-            return response()->json(['Message' => 'Not allowed to edit roles: '], 500);
+            return response()->json(['message' => 'Not allowed to edit roles: '], 500);
         }
         try {
             $role = Role::find($request->id);
-            $role->name -> $request->name;
-            $role->description -> $request->description;
-            $role->meta -> $request->meta;
+            if ($role == null) {
+                return response()->json(['message' => 'Role not found. '], 404);
+            }
+            $role->name = $request->name ?? $role->name;
+            $role->description = $request->description ?? $role->description;
+            $role->meta = $request->meta ?? $role->meta;
             $role->save();
 
             DB::delete('delete from role_permissions where role=?)', [$request->id]);
-            $permissions = $request->permission;
+            $permissions = $request->permissions();
             foreach ($permissions as $permission) {
                 $rolePermission = new RolePermission([
                     'role' => $role->id,
-                    'permission' => $permission,
+                    'permission' => $permission->id,
                 ]);
                 $rolePermission->save();
             }
 
-            return response()->json(['Message' => 'Updated successfully'], 200);
+            return response()->json(['message' => 'Updated successfully'], 200);
         } catch (Exception $ex) {
-            return response()->json(['Message' => 'Could not save role: '  . $ex->getMessage()], 500);
+            return response()->json(['message' => 'Could not save role: '  . $ex->getMessage()], 500);
         }
     }
 }
